Propagated the following commit by Michael Larson from Bakersfield:
The portion of commit propagated applies only to 'dhcp.c'  The propagation is for implementation of DHCP server operational
mode commands in Glendale.

added packet offer, request counters to system--writes out to /var/log. temporarily removed pool tag from lease file and the
parsing of the lease file was barfing on the uncognized tag--which was causing a problem. might need to write this to a
seperate file for later consumption.

http://suva.vyatta.com/git/?p=dhcp.git;a=commitdiff;h=09469ed04229079d81ade8c55f0719e5ea9f7c81